Add support for kilnhg repositories.  (Only public repositories for now.)
Most of the work is in hg.py, to download the contents of a file in
order to create a diff.  I also added work to the forms code so users
can pick a kiln repository from the commandline.

This is not at all tested; we'll push it and see what happens!