Fix the kiln fields to be not-required.
I thought that 'required=True' meant that it was required only when
using a hosting service that listed it in its 'fields'.  But it turns
out it means required all the time.  Since the kiln_* fields are only
meaningful for kiln, they're definitely *not* required all the time,
so I have to mark them as optional.